test: only cat test coverage output file whene said file exists

+++ ci/travis.sh
@@ -8,10 +8,8 @@ go test -i ./...
 for package in "${PACKAGES[@]}"
 do
   go test -coverprofile=tmp.cov gobot.io/x/$package
-  if [ $? -ne 0 ]
-  then
-    EXITCODE=1
+  if [ -f tmp.cov ]; then
+     cat tmp.cov >> coverage.txt
+     rm tmp.cov
   fi
-  cat tmp.cov >> coverage.txt
-  rm tmp.cov
 done
